pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shellm"
version = "0.1.0"
description = "Minimal conversational shell prototype"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "prompt_toolkit>=3.0.43,<4",
    "python-dotenv>=1.0,<2",
]

[project.scripts]
shellm = "shellm.cli:main"

[tool.setuptools.packages.find]
include = ["shellm", "shellm.*"]

[tool.setuptools.package-data]
"shellm.prompts" = ["*.txt"]
